simplify waiting for suspension
